Refactored SecretCode class to ComputerPlayer class.
ComputerPlayer still works properly in code-breaker mode. The CompluterPlayer class will be further built out so that it will be used in either game mode. Build_code function was modified so that it can build guesses for the computer player when trying to guess the secret code from the player.